requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
altgraph @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
exceptiongroup==1.2.2
Faker==30.0.0
future @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
idna==3.10
iniconfig==2.0.0
macholib @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
requests==2.32.3
six @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
tomli==2.0.1
urllib3==2.2.3
